+++ pkg/ingester/ingester_test.go
@@ -357,7 +357,7 @@ func TestIngesterAppendBlankLabel(t *testing.T) {
 
 func TestIngesterUserSeriesLimitExceeded(t *testing.T) {
 	limits := defaultLimitsTestConfig()
-	limits.MaxSeriesPerUser = 1
+	limits.MaxLocalSeriesPerUser = 1
 
 	_, ing := newTestStore(t, defaultIngesterTestConfig(), defaultClientTestConfig(), limits)
 	defer ing.Shutdown()
@@ -414,7 +414,7 @@ func TestIngesterUserSeriesLimitExceeded(t *testing.T) {
 
 func TestIngesterMetricSeriesLimitExceeded(t *testing.T) {
 	limits := defaultLimitsTestConfig()
-	limits.MaxSeriesPerMetric = 1
+	limits.MaxLocalSeriesPerMetric = 1
 
 	_, ing := newTestStore(t, defaultIngesterTestConfig(), defaultClientTestConfig(), limits)
 	defer ing.Shutdown()