+++ pkg/util/validation/limits_test.go
@@ -1,0 +1,41 @@
+package validation
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestLimits_Validate(t *testing.T) {
+	t.Parallel()
+
+	tests := map[string]struct {
+		limits           Limits
+		shardByAllLabels bool
+		expected         error
+	}{
+		"max-global-series-per-user disabled and shard-by-all-labels=false": {
+			limits:           Limits{MaxGlobalSeriesPerUser: 0},
+			shardByAllLabels: false,
+			expected:         nil,
+		},
+		"max-global-series-per-user enabled and shard-by-all-labels=false": {
+			limits:           Limits{MaxGlobalSeriesPerUser: 1000},
+			shardByAllLabels: false,
+			expected:         errMaxGlobalSeriesPerUserValidation,
+		},
+		"max-global-series-per-user disabled and shard-by-all-labels=true": {
+			limits:           Limits{MaxGlobalSeriesPerUser: 1000},
+			shardByAllLabels: true,
+			expected:         nil,
+		},
+	}
+
+	for testName, testData := range tests {
+		testData := testData
+
+		t.Run(testName, func(t *testing.T) {
+			assert.Equal(t, testData.expected, testData.limits.Validate(testData.shardByAllLabels))
+		})
+	}
+}